apps/etl-brand/requirements.txt
typer==0.9.0
pandas==2.1.4
asyncpg==0.29.0
pyyaml==6.0.1